+++ src/esp/bindings/CMakeLists.txt
@@ -2,9 +2,11 @@ find_package(MagnumBindings REQUIRED Python)
 
 pybind11_add_module(
   habitat_sim_bindings
+  bindings.h
   bindings.cpp
   AttributesBindings.cpp
   AttributesManagersBindings.cpp
+  CoreBindings.cpp
   GeoBindings.cpp
   GfxBindings.cpp
   MetadataMediatorBindings.cpp
